<==== Mercurial data ====>
Branch: webcam_plugin

+++ func_tests/framework/os_access/ssh_path.py
@@ -34,11 +34,11 @@ def _raising_on_exit_status(exit_status_to_error_cls):
 
 class SSHPath(FileSystemPath, PurePosixPath):
     __metaclass__ = ABCMeta
-    _ssh_access = abstractproperty()
+    _ssh = abstractproperty()  # PurePath's manipulations can preserve only the type.
 
     @classmethod
     def home(cls):
-        return cls(cls._ssh_access.run_sh_script('echo ~').rstrip())
+        return cls(cls._ssh.run_sh_script('echo ~').rstrip())
 
     @classmethod
     def tmp(cls):
@@ -46,18 +46,18 @@ class SSHPath(FileSystemPath, PurePosixPath):
 
     def exists(self):
         try:
-            self._ssh_access.run_command(['test', '-e', self])
+            self._ssh.run_command(['test', '-e', self])
         except exit_status_error_cls(1):
             return False
         else:
             return True
 
-    @_raising_on_exit_status({2: NotADir, 3: NotAFile})
+    @_raising_on_exit_status({2: DoesNotExist, 3: NotAFile})
     def unlink(self):
-        self._ssh_access.run_sh_script(
+        self._ssh.run_sh_script(
             # language=Bash
             '''
-                test ! -e "$SELF" && >&2 echo "does not exist: $SELF" && echo 2
+                test ! -e "$SELF" && >&2 echo "does not exist: $SELF" && exit 2
                 test ! -f "$SELF" && >&2 echo "not a file: $SELF" && exit 3
                 rm -v -- "$SELF"
                 ''',
@@ -70,7 +70,7 @@ class SSHPath(FileSystemPath, PurePosixPath):
         if self.parts[0] == '~':
             return self.home().joinpath(*self.parts[1:])
         user_name = self.parts[0][1:]
-        output = self._ssh_access.run_command(['getent', 'passwd', user_name])
+        output = self._ssh.run_command(['getent', 'passwd', user_name])
         if not output:
             raise RuntimeError("Can't determine home directory for {!r}".format(user_name))
         user_home_dir = output.split(':')[6]
@@ -78,7 +78,7 @@ class SSHPath(FileSystemPath, PurePosixPath):
 
     @_raising_on_exit_status({2: DoesNotExist, 3: NotADir})
     def glob(self, pattern):
-        output = self._ssh_access.run_sh_script(
+        output = self._ssh.run_sh_script(
             # language=Bash
             '''
                 test ! -e "$SELF" && >&2 echo "does not exist: $SELF" && exit 2
@@ -92,7 +92,7 @@ class SSHPath(FileSystemPath, PurePosixPath):
 
     @_raising_on_exit_status({2: BadParent, 3: AlreadyExists, 4: BadParent})
     def mkdir(self, parents=False, exist_ok=False):
-        self._ssh_access.run_sh_script(
+        self._ssh.run_sh_script(
             # language=Bash
             '''
                 ancestor="$DIR"
@@ -117,7 +117,7 @@ class SSHPath(FileSystemPath, PurePosixPath):
 
     def rmtree(self, ignore_errors=False):
         try:
-            self._ssh_access.run_sh_script(
+            self._ssh.run_sh_script(
                 # language=Bash
                 '''
                     test -e "$SELF" || exit 2
@@ -134,7 +134,7 @@ class SSHPath(FileSystemPath, PurePosixPath):
 
     @_raising_on_exit_status({2: DoesNotExist, 3: NotAFile})
     def read_bytes(self):
-        return self._ssh_access.run_sh_script(
+        return self._ssh.run_sh_script(
             # language=Bash
             '''
                 test ! -e "$SELF" && >&2 echo "does not exist: $SELF" && exit 2
@@ -145,7 +145,7 @@ class SSHPath(FileSystemPath, PurePosixPath):
 
     @_raising_on_exit_status({2: BadParent, 3: BadParent, 4: NotAFile})
     def write_bytes(self, contents):
-        output = self._ssh_access.run_sh_script(
+        output = self._ssh.run_sh_script(
             # language=Bash
             '''
                 parent="$(dirname "$SELF")"
@@ -167,3 +167,11 @@ class SSHPath(FileSystemPath, PurePosixPath):
     def write_text(self, data, encoding='ascii', errors='strict'):
         # ASCII encoding is single used encoding in the project.
         self.write_bytes(data.encode(encoding=encoding, errors=errors))
+
+
+def make_ssh_path_cls(ssh):
+    """Separate function to be used within SSHAccess and with ad-hoc SSH"""
+    class SpecificSSHPath(SSHPath):
+        _ssh = ssh
+
+    return SpecificSSHPath
